+++ packages/iobio-react-components/src/index.ts
@@ -38,7 +38,7 @@ export {
 	type BamPercentKey,
 } from './utils/constants';
 
-export { getBooleanAttributes, isOutlierKey, setElementStyles } from './utils';
+export { getBooleanAttributes, isOutlierKey, setElementStyles } from './utils/functions';
 
 // Init Iobio Charts
 // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/import#import_a_module_for_its_side_effects_only
